rename modal to layout
